chore: clarify runtime mode and event boundaries
