+++ src/web_interface/control/events.rs
@@ -33,7 +33,7 @@ pub async fn history(state: Data<SharedState>, query: Query<EventHistoryQuery>) 
 	let mut rows: Vec<_> = h
 		.iter()
 		.rev()
-		.filter(|item| kind.map(|k| control_event_kind(&item.event) == k).unwrap_or(true))
+		.filter(|item| kind.map(|k| item.event.kind() == k).unwrap_or(true))
 		.take(limit)
 		.cloned()
 		.collect();
@@ -42,22 +42,4 @@ pub async fn history(state: Data<SharedState>, query: Query<EventHistoryQuery>) 
 		"events": rows,
 		"limit": limit,
 	}))
-}
-
-fn control_event_kind(ev: &ControlEvent) -> &'static str {
-	match ev {
-		ControlEvent::ChannelDatum { .. } => "channel_datum",
-		ControlEvent::Lagged { .. } => "lagged",
-		ControlEvent::Heartbeat { .. } => "heartbeat",
-		ControlEvent::PauseState { .. } => "pause_state",
-		ControlEvent::Reloaded { .. } => "reloaded",
-		ControlEvent::ProcessorInvoked { .. } => "processor_invoked",
-		ControlEvent::Restarting { .. } => "restarting",
-		ControlEvent::FlowgraphReloaded { .. } => "flowgraph_reloaded",
-		ControlEvent::ManagedAppState { .. } => "managed_app_state",
-		ControlEvent::RestartRecommended { .. } => "restart_recommended",
-		ControlEvent::RuntimeModeChanged { .. } => "runtime_mode_changed",
-		ControlEvent::RuntimeModeManagedApps { .. } => "runtime_mode_managed_apps",
-		ControlEvent::OAuthStatus { .. } => "oauth_status",
-	}
 }
